requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
bitarray==2.6.0
certifi==2022.9.24
charset-normalizer==2.1.1
contextlib2==21.6.0
cytoolz==0.12.0
discord==2.0.0
discord.py==2.0.1
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.1
hexbytes==0.3.0
idna==3.4
ipfshttpclient==0.8.0a2
jsonschema==4.16.0
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.19.5
pycodestyle==2.9.1
pycryptodome==3.15.0
pyrsistent==0.18.1
PyYAML==6.0
requests==2.28.1
rlp==2.0.1
schema==0.7.5
six==1.16.0
toml==0.10.2
toolz==0.12.0
urllib3==1.26.12
varint==1.0.2
web3==5.31.1
websockets==9.1
yarl==1.8.1
